full_requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
conllu==6.0.0
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
en_core_sci_sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz#sha256=a54a114ee11dab7cf6db8cf69ef485145011537f45f3c16c71338c48fd6f070b
en_ner_bc5cdr_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz#sha256=ffc73130a710edf851206199720cb2c744a043e032f5da6ba4bb36863deca778
executing==2.2.0
fonttools==4.59.1
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.4.19
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
medspacy==1.3.1
medspacy_quickumls==3.2
medspacy_unqlite==0.9.8
murmurhash==1.0.13
narwhals==2.2.0
nest-asyncio==1.6.0
nltk==3.9.1
nmslib-metabrainz==2.1.3
numpy==1.26.4
orjson==3.11.3
packaging==24.2
pandas==2.2.2
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
phrasemachine==1.0.7
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.6.0
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pybind11==3.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyFastNER==1.0.10
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.0
PyRuSH==1.0.10
pysbd==0.3.4
pysimstring==1.3.0
pytesseract==0.3.10
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
quicksectx==0.4.1
referencing==0.36.2
regex==2024.5.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.0
scikit-learn==1.5.1
scipy==1.16.1
scispacy==0.5.5
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
streamlit==1.37.1
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.66.5
traitlets==5.14.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
wheel==0.45.1
wrapt==1.17.3
zstandard==0.24.0
